Fixed turn Fixed walls and wheel detectors

+++ RoombaSimulator/src/org/github/triman/roomba/simulator/physics/PhysicalRoomba.scala
@@ -9,7 +9,7 @@ import Math._
 object PhysicalRoomba {	
 	val rightBumper = 0 to 4 map(i => new Point2D.Double(175*sin(PI/10*i), 175*cos(PI/10*i)))	// 5 points on the right side
 	val leftBumper	= 5 to 9 map(i => new Point2D.Double(175*sin(PI/10*i), 175*cos(PI/10*i))) // 5 points on the left side
-	val wall				= new Point2D.Double(0, 178)	// 10mm outside on the right, on the horizontal axis
+	val wall				= new Point2D.Double(0, 195)	// 10mm outside on the right, on the horizontal axis
 	val backBumper 	= 10 to 19 map(i => new Point2D.Double(175*sin(PI/10*i), 175*cos(PI/10*i)))	// 10 virtual back bumpers (not present on the roomba)
 	
 	val casterWheel = new Point2D.Double(130, 0)	// on the axis, 130mm from the center
